refactor: 엔티티 개행 및 인수 테스트 helper 적용
