+++ src/test/java/integration/product/ProductIntegrationTest.java
@@ -1,42 +1,21 @@
-package integration;
+package integration.product;
 
 import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
 
 import com.woowa.woowakit.domain.product.dto.request.ProductCreateRequest;
 import com.woowa.woowakit.domain.product.dto.response.ProductDetailResponse;
 
-import io.restassured.RestAssured;
+import integration.IntegrationTest;
+import integration.helper.CommonRestAssuredUtils;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 
 @DisplayName("Product 인수 테스트")
 public class ProductIntegrationTest extends IntegrationTest {
-
-	// todo: Helper 클래스 패키지, Test Fixture 패키지 어디에 둘 지 토론
-	public static <T> ExtractableResponse<Response> post(String url, T body) {
-		return RestAssured.given().log().all()
-			.body(body)
-			.contentType(MediaType.APPLICATION_JSON_VALUE)
-			.when()
-			.post(url)
-			.then().log().all()
-			.extract();
-	}
-
-	public static <T> ExtractableResponse<Response> get(String url) {
-		return RestAssured.given().log().all()
-			.contentType(MediaType.APPLICATION_JSON_VALUE)
-			.accept(MediaType.APPLICATION_JSON_VALUE)
-			.when()
-			.get(url)
-			.then().log().all()
-			.extract();
-	}
 
 	@Test
 	@DisplayName("상품 이름, 가격, 이미지 주소를 입력해 Product를 생성할 수 있다.")
@@ -45,7 +24,7 @@ public class ProductIntegrationTest extends IntegrationTest {
 		ProductCreateRequest request = ProductCreateRequest.of("test", 3000L, "testImage");
 
 		// when
-		ExtractableResponse<Response> response = post("/products", request);
+		ExtractableResponse<Response> response = CommonRestAssuredUtils.post("/products", request);
 
 		//then
 		assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
@@ -57,10 +36,10 @@ public class ProductIntegrationTest extends IntegrationTest {
 	void findById() {
 		// given
 		ProductCreateRequest request = ProductCreateRequest.of("test", 3000L, "testImage");
-		String location = post("/products", request).header("Location");
+		String location = CommonRestAssuredUtils.post("/products", request).header("Location");
 
 		// when
-		ExtractableResponse<Response> response = get(location);
+		ExtractableResponse<Response> response = CommonRestAssuredUtils.get(location);
 
 		// then
 		assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
